# swagger.yaml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0
  description: API Documentation
servers:
  - url: http://localhost:8080
    description: HTTP Server
...
paths:
...
  /profile:
    get:
      tags:
        - Profile
      summary: Получить страницу профиля
      description: Возвращает страницу профиля мастера.
      responses:
        '200':
          description: Успешный ответ
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Мастер не найден
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...
  /profile/services:
    get:
      tags:
        - Profile
      summary: Получить страницу услуг
      description: Возвращает страницу с услугами мастера.
      responses:
        '200':
          description: Успешный ответ
          content:
            text/html:
              schema:
                type: string
        '403':
          description: Необходима регистрация
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Мастер не найден
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...